refactor move button inside card
